refactor: move all interfaces to types/types.ts
